Move another test case into the integration tests
